docs(mft): complete the forensic_flags bit table; split index/convert.rs out
The comment above the forensic-flag accessors listed only bits 0-2 and
stopped, which reads as "bit 3 onwards is free". It is not: bits 3-5 are
has_default_data, has_i30_stream and is_unified, written by their own
setters, which is why `set_forensic_flags` masks them through with
`& 0b11_1000`. Bit 3 feeds `compute_tree_metrics`, where it separates
"empty $DATA" from "no $DATA", so a new flag landing there would surface
as wrong treesize/tree_allocated far from the cause. Only bits 6-7 are
free. The table is now complete and says so.

types.rs sat at exactly 800 LOC, so the full comment pushed it over the
policy limit. Rather than trim the docs to fit or take a size exception,
lift the "Safe Casting Helpers" section — 15 numeric conversions with no
dependency on the record structs, and no callers inside types.rs — into a
new `index/convert.rs` (211 LOC), leaving types.rs at 610 with room to
document itself. Re-exported from `index.rs` under the same public paths,
so no call site outside the crate changes; `index/base.rs` was the one
in-crate user of a `types::`-qualified path and now points at `convert`.
